moved personalized_loaders & loaders_with_permissions to their own file
